chore(test): restructure src/test sub-directory to match src/main
