--- src/test/java/io/cryostat/jfr/datasource/tests/NativeDatasourceIT.java
@@ -1,6 +1,0 @@
-package io.cryostat.jfr.datasource.tests;
-
-import io.quarkus.test.junit.NativeImageTest;
-
-@NativeImageTest
-public class NativeDatasourceIT extends DatasourceTest {}
