+++ src/test/java/itest/DatasourceIT.java
@@ -1,0 +1,6 @@
+package itest;
+
+import io.quarkus.test.junit.QuarkusTest;
+
+@QuarkusTest
+public class DatasourceIT {}
