+++ src/test/java/itest/NativeDatasourceIT.java
@@ -1,0 +1,6 @@
+package itest;
+
+import io.quarkus.test.junit.NativeImageTest;
+
+@NativeImageTest
+public class NativeDatasourceIT extends DatasourceIT {}
